renamed PathParser to FsParser, now it doesn't parse tag-related stuff
e.g. artist, album etc. this caused some issues when folders were
created for upload it would use the "mutagen tags" only but moving
"other files" would use the track entity which would have these "fake"
tags on them 'cause the path structure for these to be inconsitent with
the actual filesystem structure. also, this sort of goes against our
general philosophy of not "faking" stuff.

note though that there is however one instance of when we fake the
metadata. in uploads we have a fallback artist if there's no artist tag
on the file, this is a lazy-fix for making it easier to fix tags after
upload if they're missing.